requirements.txt
fastapi[standard]
numpy
pandas
tensorflow
scikit-learn
joblib
opencv-python-headless
pillow
passlib[bcrypt]
sqlmodel